tests/wpt/web-platform-tests/html/canvas/offscreen/text/2d.text.drawing.style.fontKerning.with.uppercase.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>OffscreenCanvas test: 2d.text.drawing.style.fontKerning.with.uppercase</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.text.drawing.style.fontKerning.with.uppercase</h1>
<p class="desc">Testing basic functionalities of fontKerning for canvas</p>


<script>
var t = async_test("Testing basic functionalities of fontKerning for canvas");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

var offscreenCanvas = new OffscreenCanvas(100, 50);
var ctx = offscreenCanvas.getContext('2d');

_assertSame(ctx.fontKerning, "Auto", "ctx.fontKerning", "\"Auto\"");
ctx.fontKerning = "Normal";
_assertSame(ctx.fontKerning, "Normal", "ctx.fontKerning", "\"Normal\"");
ctx.fontKerning = "Auto";
ctx.fontKerning = "normal";
_assertSame(ctx.fontKerning, "Normal", "ctx.fontKerning", "\"Normal\"");
ctx.fontKerning = "Auto";
ctx.fontKerning = "noRmal";
_assertSame(ctx.fontKerning, "Normal", "ctx.fontKerning", "\"Normal\"");
ctx.fontKerning = "Auto";
ctx.fontKerning = "NoRMal";
_assertSame(ctx.fontKerning, "Normal", "ctx.fontKerning", "\"Normal\"");
ctx.fontKerning = "Auto";
ctx.fontKerning = "NORMAL";
_assertSame(ctx.fontKerning, "Normal", "ctx.fontKerning", "\"Normal\"");

ctx.fontKerning = "None";
_assertSame(ctx.fontKerning, "None", "ctx.fontKerning", "\"None\"");
ctx.fontKerning = "Auto";
ctx.fontKerning = "none";
_assertSame(ctx.fontKerning, "None", "ctx.fontKerning", "\"None\"");
ctx.fontKerning = "Auto";
ctx.fontKerning = "nOne";
_assertSame(ctx.fontKerning, "None", "ctx.fontKerning", "\"None\"");
ctx.fontKerning = "Auto";
ctx.fontKerning = "nonE";
_assertSame(ctx.fontKerning, "None", "ctx.fontKerning", "\"None\"");
ctx.fontKerning = "Auto";
ctx.fontKerning = "NONE";
_assertSame(ctx.fontKerning, "None", "ctx.fontKerning", "\"None\"");
t.done();

});
</script>
